Avoid some signed/unsigned warnings

+++ dune/gfe/localgeodesicfeadolcstiffness.hh
@@ -254,7 +254,7 @@ assembleGradientAndHessian(const Entity& element,
 
     std::fill(v.begin(), v.end(), 0.0);
 
-    for (int i=0; i<nDofs; i++)
+    for (size_t i=0; i<nDofs; i++)
       for (int ii=0; ii<blocksize; ii++)
       {
         // Evaluate Hessian in the direction of each vector of the orthonormal frame
@@ -266,7 +266,7 @@ assembleGradientAndHessian(const Entity& element,
         if (rc < 0)
           DUNE_THROW(Dune::Exception, "ADOL-C has returned with error code " << rc << "!");
 
-        for (int j=0; j<nDoubles; j++)
+        for (size_t j=0; j<nDoubles; j++)
           embeddedHessian[i][j/embeddedBlocksize][ii][j%embeddedBlocksize] = w[j];
 
         // Make v the null vector again
